moved types query building to utils file
